backend/requirements.txt
langgraph
langchain
pydantic
sqlalchemy
psycopg2-binary
pandas
fastmcp
openai
anthropic
python-dotenv
fastapi
uvicorn
networkx
matplotlib
scipy
requests
